requirements.txt
pyrogram
tgcrypto
aiohttp
aiofiles
tortoise-orm
bs4